refactor: modularize report generation node
